feat: don't create a new signal unless a user clicks save
